WIP: Discovery, Monitoring, Security
* TLS in gRPC: wip
* monitoring & discovery
  * http /status: json object with stats
  * http /info: description of the service
  * http /health: expected behavior by Hashicorp Consul
  * grpc equivalents
