+++ .gitignore
@@ -1,10 +1,3 @@
-pkg/blobindex-proto/blobindex.pb.go
-./object-storage
-./blob-client
-./blob-server
-./blobindex-client
-./blobindex-server
-
 # Jetbrains IDE
 cmake-build-*/
 .idea/
