+++ pkg/gunkan-index-proto/.gitignore
@@ -1,0 +1,1 @@
+index.pb.go
